+++ public/views/section/header/default.php
@@ -1,0 +1,6 @@
+<section id="home-intro" class="site-intro">
+    <div class="intro-header">
+        <h1 class="intro-header-title"><?php echo esc_html( get_theme_mod( 'theme_home_header_title' , 'Our Mission Goes Here!' ) ); ?></h1>
+        <div class="intro-header-description"><?php echo esc_html( get_theme_mod( 'theme_home_header_description' , 'Your description goes here!' ) ); ?></div>
+    </div>
+</section>