+++ public/views/sidebar/portfolio.php
@@ -1,0 +1,6 @@
+<div id="aside" class="widget-area">
+	<aside id="post-thumbnail" class="widget widget_post_thumbnail">
+		<?php the_post_thumbnail(); ?>
+	</aside>
+	<?php dynamic_sidebar( $data->location ); ?>
+</div>